+++ packages/docs-a/gatsby-config.js
@@ -1,8 +1,8 @@
-const config = require("./data/SiteConfig");
-const iconsConfig = require("./static/icons.json");
-const urljoin = require("url-join");
+const config = require('./data/SiteConfig')
+const iconsConfig = require('./static/icons.json')
+const urljoin = require('url-join')
 
-const regexExcludeRobots = /^(?!\/(dev-404-page|404|offline-plugin-app-shell-fallback|tags|categories)).*$/;
+const regexExcludeRobots = /^(?!\/(dev-404-page|404|offline-plugin-app-shell-fallback|tags|categories)).*$/
 
 require('dotenv').config({
   path: `.env.${process.env.NODE_ENV}`,
@@ -33,69 +33,85 @@ const allDocsPageQuery = `{
       }	
     }	
   }
-}`;
+}`
 
 const queries = [
   {
     query: allDocsPageQuery,
-    transformer: ({ data }) => data.allMarkdownRemark.edges.map(({ node }) => node), // optional    
+    transformer: ({ data }) =>
+      data.allMarkdownRemark.edges.map(({ node }) => node), // optional
     indexName: 'docs-', // overrides main index name, optional
   },
-];
-
+]
 
 module.exports = {
   pathPrefix: config.pathPrefix,
   siteMetadata: {
     siteUrl: urljoin(config.siteUrl, config.pathPrefix),
-    title: config.siteTitle
+    title: config.siteTitle,
   },
   plugins: [
     'gatsby-plugin-typescript',
+    // {
+    //   resolve: 'gatsby-plugin-module-resolver',
+    //   options: {
+    //     root: './', // <- will be used as a root dir
+    //     alias: {
+    //       '@m00n/gatsby-docs-ui': '../packages/gatsby-docs-ui/src',
+    //       '@m00n/gatsby-plugin-nav-tree':
+    //         '../packages/gatsby-plugin-nav-tree/src',
+    //     },
+    //   },
+    // },
     {
-      resolve: "@m00n/gatsby-plugin-nav-tree",
+      resolve: '@m00n/gatsby-plugin-nav-tree',
       options: {
-        ignorePaths: ["/categories", "/tags",  "/404", "/offline-plugin-app-shell-fallback",
-        "/dev-404-page"]
-      }
+        ignorePaths: [
+          '/categories',
+          '/tags',
+          '/404',
+          '/offline-plugin-app-shell-fallback',
+          '/dev-404-page',
+        ],
+      },
     },
-    "gatsby-plugin-react-helmet",
-    "gatsby-plugin-lodash",
+    'gatsby-plugin-react-helmet',
+    'gatsby-plugin-lodash',
     {
-      resolve: "gatsby-source-filesystem",
+      resolve: 'gatsby-source-filesystem',
       options: {
-        name: "docs",
-        path: `${__dirname}/content`
-      }
+        name: 'docs',
+        path: `${__dirname}/content`,
+      },
     },
     {
-      resolve: "gatsby-transformer-remark",
+      resolve: 'gatsby-transformer-remark',
       options: {
         plugins: [
           {
             resolve: `gatsby-remark-sequence`,
             options: {
-              theme: 'simple'
-            }
-          },          
+              theme: 'simple',
+            },
+          },
           {
-            resolve: "gatsby-remark-embed-youtube",
+            resolve: 'gatsby-remark-embed-youtube',
             options: {
               width: 1000,
-              height: 400
-            }
+              height: 400,
+            },
           },
           {
-            resolve: "gatsby-remark-relative-images"
+            resolve: 'gatsby-remark-relative-images',
           },
           {
-            resolve: "gatsby-remark-images",
+            resolve: 'gatsby-remark-images',
             options: {
-              maxWidth: 672
-            }
-          },         
+              maxWidth: 672,
+            },
+          },
           {
-            resolve: "gatsby-remark-responsive-iframe"
+            resolve: 'gatsby-remark-responsive-iframe',
           },
           {
             resolve: `gatsby-remark-prismjs`,
@@ -107,7 +123,7 @@ module.exports = {
               // you may use this to prevent Prism from re-processing syntax.
               // This is an uncommon use-case though;
               // If you're unsure, it's best to use the default value.
-              classPrefix: "language-",
+              classPrefix: 'language-',
               // This is used to allow setting a language for inline code
               // (i.e. single backticks) by creating a separator.
               // This separator is a string and will do no white-space
@@ -133,29 +149,29 @@ module.exports = {
               noInlineHighlight: false,
             },
           },
-          "gatsby-remark-copy-linked-files",
-          "gatsby-remark-autolink-headers"
-        ]
-      }
-    },   
-    "gatsby-transformer-sharp",
-    "gatsby-plugin-sharp",
+          'gatsby-remark-copy-linked-files',
+          'gatsby-remark-autolink-headers',
+        ],
+      },
+    },
+    'gatsby-transformer-sharp',
+    'gatsby-plugin-sharp',
     {
-      resolve: "gatsby-plugin-manifest",
+      resolve: 'gatsby-plugin-manifest',
       options: {
         name: config.siteTitle,
         short_name: config.siteTitleShort,
         description: config.siteDescription,
         start_url: config.pathPrefix,
-        background_color: "#334058",
-        theme_color: "#6ec5ff",
-        display: "standalone",
-        icons: iconsConfig.icons
-      }
+        background_color: '#334058',
+        theme_color: '#6ec5ff',
+        display: 'standalone',
+        icons: iconsConfig.icons,
+      },
     },
     'gatsby-plugin-offline',
     {
-      resolve: "gatsby-plugin-netlify-cms",
+      resolve: 'gatsby-plugin-netlify-cms',
       options: {
         modulePath: `${__dirname}/src/cms/cms.js`,
       },
@@ -173,5 +189,5 @@ module.exports = {
     //   },
     // },
     // make sure to put last in the array
-  ]
-};
+  ],
+}