+++ tmp/gatsby-starter-docs-material-ui/data/SiteConfig.js
@@ -1,16 +1,17 @@
 module.exports = {
   siteTitle: '****** DOCS', // Site title.
-  siteTitleShort: '6star DOCS', // Short site title for homescreen (PWA). Preferably should be under 12 characters to prevent truncation.  
+  siteTitleShort: '6star DOCS', // Short site title for homescreen (PWA). Preferably should be under 12 characters to prevent truncation.
   siteTitleAlt: '****** DOCS', // Alternative site title for SEO.
-  siteDescription: "for ******",
-  siteLogo: 'https://s.gravatar.com/avatar/cb868bd3f76af9631aa55aca22615673?size=1024', // Logo used for SEO and manifest.
+  siteDescription: 'for ******',
+  siteLogo:
+    'https://s.gravatar.com/avatar/cb868bd3f76af9631aa55aca22615673?size=1024', // Logo used for SEO and manifest.
   siteUrl: 'https://6stars.netlify.com', // Domain of your website without pathPrefix.
   siteRss: '/rss.xml', // Path to the RSS file.
-  pathPrefix: '/', // Prefixes all links.
+  pathPrefix: '/docs', // Prefixes all links.
   siteGATrackingID: '', // Tracking code ID for google analytics.
   defaultCategoryID: 'api', // Default category for docs.
   dateFromFormat: 'MM-DD-YYYY', // Date format used in the frontmatter.
   dateFormat: 'MM/DD/YYYY', // Date format for display.
   userName: '****** User', // Username to display in the author segment.
-  copyright: 'Copyright © 2018. ******', // Copyright string for the footer of the website and RSS feed.
-};
+  copyright: 'Copyright © 2019. ******', // Copyright string for the footer of the website and RSS feed.
+}