+++ packages/docs-a/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.3](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.0.2...@m00n/docs-a@1.0.3) (2019-02-11)
+
+**Note:** Version bump only for package @m00n/docs-a
+
+## [1.0.2](https://github.com/6stars/gatsby-docs/compare/@m00n/docs-a@1.0.1...@m00n/docs-a@1.0.2) (2019-02-11)
+
+**Note:** Version bump only for package @m00n/docs-a
+
 ## 1.0.1 (2019-02-09)
 
 ### Bug Fixes