+++ packages/gatsby-docs-ui/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.3](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.1.2...@m00n/gatsby-docs-ui@1.1.3) (2019-02-11)
+
+### Bug Fixes
+
+- module-resolver babel ([13ac272](https://github.com/6stars/gatsby-docs/commit/13ac272))
+
+## [1.1.2](https://github.com/6stars/gatsby-docs/compare/@m00n/gatsby-docs-ui@1.1.1...@m00n/gatsby-docs-ui@1.1.2) (2019-02-11)
+
+**Note:** Version bump only for package @m00n/gatsby-docs-ui
+
 ## 1.1.1 (2019-02-09)
 
 **Note:** Version bump only for package @m00n/gatsby-docs-ui